Moved TradeWidget to apps/gp-v1/components
