+++ src/apps/gp-v1/components/TradeWidget/SwapIcon.tsx
