+++ src/apps/gp-v1/components/TradeWidget/TradeWidget.styled.ts
