+++ src/apps/gp-v1/components/TradeWidget/TxMessage.tsx
