+++ src/apps/gp-v1/components/TradeWidget/utils.ts
