+++ src/components/trade/Price.tsx
@@ -16,8 +16,8 @@ import { FormInputError } from 'components/common/FormInputError'
 import { FormMessage } from 'components/common/FormMessage'
 
 // TradeWidget: subcomponents
-import { TradeFormData } from 'components/TradeWidget'
-import { useNumberInput } from 'components/TradeWidget/useNumberInput'
+import { TradeFormData } from 'apps/gp-v1/components/TradeWidget'
+import { useNumberInput } from 'apps/gp-v1/components/TradeWidget/useNumberInput'
 import { SwapPrice } from 'components/common/SwapPrice'
 
 const Wrapper = styled.div`
