+++ src/hooks/usePlaceOrder.ts
@@ -11,7 +11,7 @@ import { logDebug, formatDistanceStrict, formatDateLocaleShortTime, batchIdToDat
 import { txOptionalParams as defaultTxOptionalParams } from 'utils/transaction'
 import { BATCHES_TO_WAIT } from 'const'
 import useSafeState from './useSafeState'
-import { getNumberOfBatchesLeftUntilNow } from 'components/TradeWidget/OrderValidity'
+import { getNumberOfBatchesLeftUntilNow } from 'apps/gp-v1/components/TradeWidget/OrderValidity'
 
 interface ConnectionParams {
   userAddress: string
